chore: upgrade Angular CLI, Angular core and PrimeNG libraries
