+++ package.json
@@ -50,15 +50,15 @@
     "generate:ch": "node ./dev/tools/contextual-help-generator/generate.js"
   },
   "dependencies": {
-    "@angular/animations": "~6.0.5",
-    "@angular/common": "~6.0.5",
-    "@angular/compiler": "~6.0.5",
-    "@angular/core": "~6.0.5",
-    "@angular/forms": "~6.0.5",
-    "@angular/http": "~6.0.5",
-    "@angular/platform-browser": "~6.0.5",
-    "@angular/platform-browser-dynamic": "~6.0.5",
-    "@angular/router": "~6.0.5",
+    "@angular/animations": "~6.1.6",
+    "@angular/common": "~6.1.6",
+    "@angular/compiler": "~6.1.6",
+    "@angular/core": "~6.1.6",
+    "@angular/forms": "~6.1.6",
+    "@angular/http": "~6.1.6",
+    "@angular/platform-browser": "~6.1.6",
+    "@angular/platform-browser-dynamic": "~6.1.6",
+    "@angular/router": "~6.1.6",
     "@kaltura-ng/kaltura-common": "~7.0.1",
     "@kaltura-ng/kaltura-logger": "~3.0.0",
     "@kaltura-ng/kaltura-primeng-ui": "~3.0.3",
@@ -79,7 +79,7 @@
     "ngx-webstorage": "~2.0.1",
     "normalize.css": "~8.0.0",
     "primeicons": "~1.0.0-beta.8",
-    "primeng": "~6.0.0-rc.1",
+    "primeng": "~6.1.3",
     "ramda": "~0.25.0",
     "rxjs": "~6.2.1",
     "rxjs-compat": "^6.0.0-rc.0",
@@ -91,9 +91,9 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "~6.0.8",
-    "@angular/compiler-cli": "~6.0.5",
-    "@angular/language-service": "~6.0.5",
+    "@angular/cli": "~6.1.5",
+    "@angular/compiler-cli": "~6.1.6",
+    "@angular/language-service": "~6.1.6",
     "@kaltura-ng/dev-workspace": "~2.4.1",
     "@types/core-js": "0.9.28",
     "@types/jquery": "^2.0.40",
@@ -105,7 +105,7 @@
     "minimist": "~1.2.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "2.7.2",
+    "typescript": "2.9.2",
     "webpack-bundle-analyzer": "~2.10.0"
   },
   "engines": {
