+++ src/applications/administration-roles-app/roles-table/roles-table.component.html
@@ -63,5 +63,5 @@
           </div>
       </ng-template>
   </p-table>
-  <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+  <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
 </div>
