+++ src/applications/administration-users-app/users/users-table.component.html
@@ -82,7 +82,7 @@
                 </tr>
             </ng-template>
         </p-table>
-        <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'"></p-menu>
+        <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'"></p-menu>
     </div>
 
 </k-area-blocker>
