+++ src/applications/content-bulk-log-app/bulk-log-table/bulk-log-table.component.html
@@ -110,5 +110,5 @@
       </ng-template>
     </p-table>
 
-    <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+    <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
   </div>
