+++ src/applications/content-categories-app/categories/categories-table/categories-table.component.html
@@ -105,5 +105,6 @@
       </ng-template>
     </p-table>
 
-    <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
   </div>
+
+  <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
