+++ src/applications/content-categories-app/category/category-metadata/category-metadata.component.html
@@ -1,6 +1,6 @@
 <k-area-blocker [showLoader]="_widgetService.showSectionLoader" [message]="_widgetService.sectionBlockerMessage">
   <div class="jumpTo">
-    <p-menu #menu popup="popup" [model]="_jumpToMenu" [appendTo]="'body'" kMenuCloseOnScroll [style]="{width: '220px'}"></p-menu>
+    <p-menu #menu [popup]="true" [model]="_jumpToMenu" [appendTo]="'body'" kMenuCloseOnScroll [style]="{width: '220px'}"></p-menu>
     <button type="button" class="kButtonDefault" pButton icon="kIcondropdown_arrow_bottom" iconPos="right" label="{{'applications.content.categoryDetails.metadata.jumpTo' | translate}}"
       (click)="menu.toggle($event)"></button>
   </div>
