+++ src/applications/content-categories-app/category/category-subcategories/category-subcategories-table/category-subcategories-table.component.html
@@ -53,6 +53,6 @@
                 <div class="emptymessage">{{_emptyMessage}}</div>
             </ng-template>
         </p-table>
-        <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+        <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
     </k-area-blocker>
 </div>
