+++ src/applications/content-drop-folders-app/drop-folders-table/drop-folders-table.component.html
@@ -120,7 +120,7 @@
         </ng-template>
       </p-table>
 
-      <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" styleClass="kActionsMenu"
+      <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" styleClass="kActionsMenu"
               kMenuCloseOnScroll></p-menu>
     </div>
   </div>
