+++ src/applications/content-entries-app/entry/entry-metadata/entry-metadata.component.html
@@ -1,7 +1,7 @@
 <k-area-blocker [showLoader]="_widgetService.showSectionLoader" [message]="_widgetService.sectionBlockerMessage"
                 [bodyScroll]="true">
   <div class="jumpTo" *ngIf="_jumpToMenu.length > 2">
-    <p-menu #menu popup="popup" [model]="_jumpToMenu" [appendTo]="'body'" kMenuCloseOnScroll [style]="{'width': '220px', 'overflow-x': 'hidden'}"></p-menu>
+    <p-menu #menu [popup]="true" [model]="_jumpToMenu" [appendTo]="'body'" kMenuCloseOnScroll [style]="{'width': '220px', 'overflow-x': 'hidden'}"></p-menu>
     <button type="button" class="kButtonDefault" pButton icon="kIcondropdown_arrow_bottom" iconPos="right"
             label="{{'applications.content.entryDetails.metadata.jumpTo' | translate}}"
             (click)="menu.toggle($event)"></button>
