+++ src/applications/content-entries-app/entry/entry-related/entry-related.component.html
@@ -82,10 +82,10 @@
                     <div class="emptymessage">{{'applications.content.table.noResults' | translate}}</div>
                 </ng-template>
             </p-table>
-            <p-menu #actionsmenu popup="popup" [model]="_actions" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
         </div>
     </div>
     <kPopupWidget #editPopup data-aid="editRelatedPopup" [popupWidth]="380" [popupHeight]="300" [modal]="true">
         <kEntryRelatedEdit [currentFile]="_currentFile" [parentPopupWidget]="editPopup"></kEntryRelatedEdit>
     </kPopupWidget>
 </k-area-blocker>
+<p-menu #actionsmenu [popup]="true" [model]="_actions" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
