+++ src/applications/content-playlists-app/playlist/playlist-content/manual/playlist-entries-table/playlist-entries-table.component.html
@@ -81,6 +81,6 @@
                 <div class="emptymessage">{{_emptyMessage}}</div>
             </ng-template>
         </p-table>
-        <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+        <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
     </div>
 </k-area-blocker>
