+++ src/applications/content-playlists-app/playlist/playlist-content/manual/playlist-entries-table/playlist-entries-table.component.ts
@@ -125,7 +125,6 @@ export class PlaylistEntriesTableComponent implements AfterViewInit, OnInit, OnD
     if (this.actionsMenu) {
       this._buildMenu(rowIndex, entry);
       this.actionsMenu.toggle(event);
-      this.actionsMenu.show(event);
     }
   }
