+++ src/applications/content-playlists-app/playlist/playlist-content/rule-based/playlist-rules-table/playlist-rules-table.component.html
@@ -61,6 +61,6 @@
                 <div class="emptymessage">{{_emptyMessage}}</div>
             </ng-template>
         </p-table>
-        <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+        <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
     </div>
 </k-area-blocker>
