+++ src/applications/content-playlists-app/playlist/playlist-content/rule-based/playlist-rules-table/playlist-rules-table.component.ts
@@ -98,7 +98,6 @@ export class PlaylistRulesTableComponent implements AfterViewInit, OnInit, OnDes
     if (this.actionsMenu) {
       this._buildMenu(rowIndex, rule);
       this.actionsMenu.toggle(event);
-      this.actionsMenu.show(event);
     }
   }
