+++ src/applications/content-playlists-app/playlists/playlists-table/playlists-table.component.html
@@ -103,5 +103,7 @@
       </ng-template>
     </p-table>
 
-    <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
   </div>
+
+  <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+
