+++ src/applications/content-syndication-app/feeds/feeds-table/feeds-table.component.html
@@ -123,5 +123,5 @@
       </ng-template>
     </p-table>
 
-    <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+    <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
   </div>
