+++ src/applications/settings-access-control-app/profiles/profiles-table/profiles-table.component.html
@@ -163,5 +163,7 @@
             </tr>
         </ng-template>
     </p-table>
-    <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
 </div>
+
+<p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+
