+++ src/applications/settings-custom-data-app/schemas/schemas-table/schemas-table.component.html
@@ -102,5 +102,6 @@
             </div>
         </ng-template>
     </p-table>
-    <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
 </div>
+
+<p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
