+++ src/applications/settings-transcoding-settings-app/transcoding-profiles/transcoding-profiles-table/transcoding-profiles-table.component.html
@@ -86,5 +86,5 @@
             </div>
         </ng-template>
     </p-table>
-    <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
+    <p-menu #actionsmenu [popup]="true" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
 </div>
