+++ src/shared/content-shared/entries/entries-table/entries-table.component.html
@@ -209,6 +209,5 @@
             </div>
         </ng-template>
     </p-table>
-
-    <p-menu #actionsmenu popup="popup" [model]="_items" [appendTo]="'body'" kMenuCloseOnScroll></p-menu>
 </div>
+
